requirements.txt
python-dotenv
textblob
tweepy
pylama
yapf
pandas
nltk